Eliminate all warnings, enforce -Werror on deprecations
Remove -Wconf:cat=deprecation:info from compiler flags so deprecation
warnings are now treated as errors like all other warnings.

- Remove 29 stale @nowarn annotations (CaseInsensitiveIntMap, FreetypeFontLoader)
- Delete all @deprecated API with no external users: 126 MeshPartBuilder/
  MeshBuilder shape methods, buffer hierarchy getters, Graphics.getRawDeltaTime,
  ScreenUtils.getFrameBufferPixmap, SpotLightEx.set, HorizontalFlowGroup,
  VerticalFlowGroup, FastAPNG, and extension deprecated methods
- Migrate ModelBuilder to use ShapeBuilder objects directly
- Fix implicit int2float conversions with explicit .toFloat
- Add proper @nowarn for JDK 20+ URL constructor and HashMap inheritance
- Mark LittleEndianInputStream.readLine as @deprecated (Java API, must keep)

+++ sge-extension/tools/src/main/scala/sge/tools/texturepacker/MaxRectsPacker.scala
@@ -104,8 +104,8 @@ class MaxRectsPacker(val settings: Settings) extends Packer {
   private def packPage(inputRects: ArrayBuffer[Rect]): Page = {
     val paddingX = settings.paddingX
     val paddingY = settings.paddingY
-    var maxWidth:  Float = settings.maxWidth
-    var maxHeight: Float = settings.maxHeight
+    var maxWidth:  Float = settings.maxWidth.toFloat
+    var maxHeight: Float = settings.maxHeight.toFloat
     var edgePadX = false
     var edgePadY = false
     if (settings.edgePadding) {
